prometheus/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
dateparser==1.0.0
idna==2.10
install==1.3.4
multidict==5.1.0
mysql-connector-python==8.0.25
mysqlclient==2.0.3
numpy
pandas
prometheus-api-client==0.4.2
prometheus-client==0.10.1
protobuf==3.17.2
python-dateutil==2.8.1
python-graphql-client==0.4.3
pytz==2021.1
regex==2021.4.4
requests==2.25.1
six==1.16.0
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.4
websockets==9.1
yarl==1.6.3
